Rename `context::internal` to `context::private`.
