+++ src/types/binary.rs
@@ -2,7 +2,7 @@
 
 use crate::borrow::private::Pointer;
 use crate::borrow::{Borrow, BorrowMut, LoanError, Ref, RefMut};
-use crate::context::internal::Env;
+use crate::context::private::Env;
 use crate::context::{Context, Lock};
 #[cfg(feature = "napi-1")]
 use crate::handle::Handle;
